backend/pyproject.toml
[project]
name = "proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "ipython>=9.6.0",
    "langchain>=1.0.0",
    "langchain-community>=0.4",
    "langchain-core>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-qdrant>=1.0.0",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "uvicorn[standard]>=0.24.0",
]
